Additional fixes to support ffxiv patch 5.0.

+++ Machina.FFXIV/Headers/Server_UpdateHpMpTp.cs
@@ -22,11 +22,13 @@ namespace Machina.FFXIV.Headers
     [StructLayout(LayoutKind.Sequential, Pack = 1)]
     public struct Server_UpdateHpMpTp
     {
+        /* Needs update for 5.0
         public Server_MessageHeader MessageHeader; // 8 DWORDS
         public UInt32 hp;
-        public UInt16 mp;
+        public UInt32 mp;
         public UInt16 tp;
         public UInt32 unknown1;
         public UInt32 unknown2;
+        */
     }
 }
